Update safejson implementation
* Remove wrapper type for Encoder and return *json.Encoder directly
* Simplify API: primary safejson functions are now just "Encoder",
  "Marshal", "marshalIndent", "Decoder" and "Unmarshal"
* Improve documentation to describe how these operations differ
  from the built-in defaults and why the differenes were chosen
