+++ safejson/marshal.go
@@ -2,11 +2,6 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// Package safejson implements encoding and decoding of JSON with the following
-// special configurations:
-//
-// - json.Decoder.UseNumber
-// - json.Encoder.SetEscapeHTML(false)
 package safejson
 
 import (
@@ -14,43 +9,27 @@ import (
 	"encoding/json"
 )
 
-// Marshal returns the JSON encoding of v.
+// Marshal returns the JSON encoding of v encoded using the "safe" encoder.
+// Unlike json.Marshal, the returned JSON bytes will not have a trailing newline.
 func Marshal(v interface{}) ([]byte, error) {
 	// go through Encoder to control SetEscapeHTML
 	var buf bytes.Buffer
-	if err := NewEncoder(&buf).Encode(v); err != nil {
+	if err := Encoder(&buf).Encode(v); err != nil {
 		return nil, err
 	}
 	return bytes.TrimSuffix(buf.Bytes(), []byte{'\n'}), nil
 }
 
-// MarshalIndent is like Marshal but indents the output to be human readable.
-func MarshalIndent(v interface{}) ([]byte, error) {
+// MarshalIndent is like Marshal but applies Indent to format the output.
+func MarshalIndent(v interface{}, prefix, indent string) ([]byte, error) {
 	b, err := Marshal(v)
 	if err != nil {
 		return nil, err
 	}
-
 	var buf bytes.Buffer
-	if err := json.Indent(&buf, b, "", "    "); err != nil {
+	err = json.Indent(&buf, b, prefix, indent)
+	if err != nil {
 		return nil, err
 	}
 	return buf.Bytes(), nil
-}
-
-// An Encoder writes JSON objects to an output stream.
-//
-// Use NewEncoder to make a new Encoder. NewEncoder is implemented differently
-// from Go version 1.7 onward.
-type Encoder struct {
-	enc *json.Encoder
-}
-
-// Encode writes the JSON encoding of v to the stream, followed by a newline
-// character.
-//
-// See the documentation for json.Marshal for details about the conversion of Go
-// values to JSON.
-func (e *Encoder) Encode(v interface{}) error {
-	return e.enc.Encode(v)
 }
